+++ mtools/util/logevent.py
@@ -96,6 +96,7 @@ class LogEvent(object):
         self._numYields = None
         self._r = None
         self._w = None
+        self._conn = None
 
         self.merge_marker_str = ''
 
@@ -291,8 +292,23 @@ class LogEvent(object):
             if match:
                 self._thread = match.group(1)
 
+            if self._thread is not None:
+                if self._thread in ['initandlisten', 'mongosMain']:
+                    if len(split_tokens) >= 5 and  split_tokens[-5][0] == '#':
+                        self._conn = 'conn' + split_tokens[-5][1:]
+                elif self._thread.startswith('conn'):
+                    self._conn = self._thread
         return self._thread
 
+    @property
+    def conn(self):
+        """
+        extract conn name if available (lazy)
+        this value is None for all lines except the log lines related to connections ,
+        that is lines matching '\[conn[0-9]+\]' or '\[(initandlisten|mongosMain)\] .* connection accepted from'
+        """
+        self.thread
+        return self._conn
 
     @property
     def operation(self):