+++ CHANGES.md
@@ -1,6 +1,11 @@
 Changes to mtools
 =================
 
+#### version 1.1.6dev
+
+  * mlogfilter: `--thread` now also matches "connection accepted" lines for that connection (#218, #219)
+
+
 #### version 1.1.5
 
   * added workaround for compile errors with XCode 5.1 / clang 3.4 (#203)